+++ database/migrations/2019_08_28_182212_add_delivery_id_to_bills_table.php
@@ -14,7 +14,7 @@ class AddDeliveryIdToBillsTable extends Migration
     public function up()
     {
         Schema::table('bills', function (Blueprint $table) {
-            $table->integer('delivery_id')->unsigned()->nullable();
+            $table->bigInteger('delivery_id')->unsigned()->nullable();
             $table->foreign('delivery_id')->references('id')->on('users');
         });
     }
